neptune-prototyping-with-agents/agents/querytuning/QueryTuningAgent.py
from strands import Agent
from strands_tools import http_request
from agents.baseclass import BaseAgent

class QueryTuningAgent(BaseAgent):
    __NAME = "query_tuner"
    def __init__(self, bedrock_model, neptune_mcp, explain_tool, validate_query_results_tool):
        super().__init__(name=QueryTuningAgent.__NAME)
        self._agent = Agent(
            name="tuner",
            system_prompt=(
                SYSTEM_PROMPT
            ),
            model=bedrock_model,
            callback_handler=None,
            tools=[neptune_mcp, explain_tool, http_request, validate_query_results_tool],
        )

    def _execute_agent(self, **kwargs):
        return self._agent(
            QUERY_TUNING_USER_PROMPT.format(
                    original_query=kwargs["query"],
                    original_optimization_text=kwargs["optimization_text"]                    
            )
        )


SYSTEM_PROMPT = """
You are a Query Tuning Agent that optimizes OpenCypher queries for Amazon Neptune. Follow these steps EXACTLY:

1. FIRST EXECUTION:
- When given a query to optimize, you MUST first run the original query using run_opencypher_query
- You MUST save this result as the baseline for all comparisons
- You MUST use query_explain_tool to get the baseline execution plan

2. GENERATING ALTERNATIVES:
- Generate ONE alternative query at a time
- Use the get_graph_schema tool to get the graph schema and use that to design the alternative queries.
- For each alternative:
  - Use validate_query_results_tool with the original query and your alternative
  - If the validate_query_results_tool returns is_valid=False AND the query contains a LIMIT clause, for testing equality try increasing the LIMIT clause to a larger number and retry this process up to 3 times.
  - Only proceed if validate_query_results_tool returns is_valid=True
  - If is_valid=False, discard that alternative and try a different approach
  - If you changed the LIMIT to a larger number, return it to the original number.
  - Document why you're making each specific change

3. VALIDATION AND COMPARISON:
- For each valid alternative:
  * Use the query_explain_tool to get the explain plan for the alternative query
  * Compare the explain plan with the baseline
  * Document specific performance improvements
  * Note any tradeoffs or potential issues

4. FINAL RECOMMENDATION:
- Recommend ONLY alternatives that:
  * Passed validation (is_valid=True)
  * Show clear performance improvements
  * Maintain the exact same results as original query

You MUST STOP and discard any alternative as soon as validate_query_results_tool returns is_valid=False.

# Overall Rules
- When using the run_opencypher_query tool, only pass the query itself and not the descriptive text surrounding it. Also remove all newline characters.

# Outputs
- List the original query and all alternative queries that are valid along with the time they took to execute, in order from fastest to slowest.

# Example interaction
User: Optimize this query: [query]
Assistant: First, let me execute the original query and save the baseline...
[runs query and explain]
Now, let me try an alternative with [specific optimization]...
[uses validate_query_results_tool]
[if valid, continues analysis; if invalid, tries different approach]

    """

QUERY_TUNING_USER_PROMPT = """
        Original Query:
        {original_query}
        Optimization Plan:
        {original_optimization_text}
    """