neptune-prototyping-with-agents/requirements.txt
strands-agents
strands-agents-tools
boto3>=1.34.0
pytest>=7.0.0
jsonschema>=4.0.0
